requirements.txt
streamlit
pymongo
anthropic
python-dotenv
requests
beautifulsoup4